+++ test/test_fibonacci_heap.c
@@ -11,7 +11,7 @@ int int_compare_fibonacci(const void *a, const void *b) {
 void test_heap_insert_and_extract_max() {
     heap *h = create_heap(FIBONACCI_HEAP,10, int_compare_fibonacci);
     int values[] = {5, 3, 10, 1, 4};
-    for (int i = 0; i < sizeof(values) / sizeof(values[0]); i++) {
+    for (int i = 0; i < (int)(sizeof(values) / sizeof(values[0])); i++) {
         h->put(h, &values[i]);
     }
     int *max = (int *)h->pop(h);
@@ -22,7 +22,7 @@ void test_heap_insert_and_extract_max() {
 void test_heap_peek_max() {
     heap *h = create_heap(FIBONACCI_HEAP, 10, int_compare_fibonacci);
     int values[] = {5, 3, 10, 1, 4};
-    for (int i = 0; i < sizeof(values) / sizeof(values[0]); i++) {
+    for (int i = 0; i < (int)(sizeof(values) / sizeof(values[0])); i++) {
         h->put(h, &values[i]);
     }
     int *max = (int *)h->pop(h);
